+++ src/hour.py
@@ -1,0 +1,16 @@
+class Hour:
+    def __init__(self, hours, minutes):
+        self.hours = hours
+        self.minutes = minutes
+
+    def __lt__(self, other):
+        if self.hours == other.hours:
+            return self.minutes < other.minutes
+
+        return self.hours < other.hours
+
+    def __le__(self, other):
+        if self.hours == other.hours:
+            return self.minutes <= other.minutes
+
+        return self.hours <= other.hours
