+++ src/main_cycle.py
@@ -1,0 +1,6 @@
+from hour import Hour
+from schedule import Schedule
+from student import Student
+
+def main_cycle(): # TODO
+    pass
