+++ src/state.py
@@ -1,37 +1,7 @@
-from copy import deepcopy   
+from copy import deepcopy
 
 def translate_subject_and_class(subject, class_number):
     return subject + str(class_number)
-
-class Hour:
-    def __init__(self, hours, minutes):
-        self.hours = hours
-        self.minutes = minutes
-
-    def __lt__(self, other):
-        if self.hours == other.hours:
-            return self.minutes < other.minutes
-
-        return self.hours < other.hours
-
-    def __le__(self, other):
-        if self.hours == other.hours:
-            return self.minutes <= other.minutes
-
-        return self.hours <= other.hours
-
-
-class Schedule:
-    def __init__(self, start_hour, end_hour):
-        self.start_hour = start_hour
-        self.end_hour = end_hour
-
-    def conflicts(self, other):
-        if other.start_hour <= self.start_hour and self.start_hour < other.end_hour:
-            return True
-        elif other.start_hour < self.end_hour and self.end_hour <= other.end_hour:
-            return True
-        else: return False
 
 class State:
     def __init__(self):
@@ -82,60 +52,3 @@ class State:
                             state_give_ins = deepcopy(self)
                             state_give_ins.trade_classes(student_id,trader_id,subject_name)
                             yield state_give_ins
-  
-
-
-
-
-class Student:
-    def __init__(self, student_id, subjects_and_classes, subject_targets, subject_give_ins, buddies):
-        self.student_id = student_id
-        self.subjects_and_classes = subjects_and_classes 
-        self.subject_targets = subject_targets
-        self.subject_give_ins = subject_give_ins
-        self.buddies = buddies
-    
-    def add_subject_and_class(self, subject_name, class_number):
-        self.subjects_and_classes[subject_name] = class_number
-
-    def add_subject_target(self, subject_name, class_number):
-        self.subject_targets[subject_name].append(class_number)
-
-    def remove_subject_target(self,subject_name):
-        del self.subject_targets[subject_name]
-
-    def add_subject_give_in(self, subject_name, class_number):
-        self.subject_give_ins[subject_name].append(class_number)
-
-    def add_buddy(self, subject_name, buddies_up):
-        for buddy in buddies_up:
-            self.buddies[subject_name] = buddy
-
-    def get_class_for_subject(self, subject):
-        if subject in self.subjects_and_classes :
-            return self.subjects_and_classes[subject]
-        return None
-
-    def set_class_for_subject(self, subject, subject_class):
-        if subject in self.subjects_and_classes:
-            self.subjects_and_classes[subject] = subject_class
-
-    def get_targets_for_subject(self, subject):
-        if subject in self.subject_targets:
-            return self.subject_targets[subject]
-        return None
-
-    def __str__(self):
-        return "\nID: " + str(self.student_id) + "\nClasses: " + str(self.subjects_and_classes) + "\nTarget Classes: " + str(self.subject_targets) + "\nGive ins" + str(self.subject_give_ins)
-         
-    def get_giveins_for_subject(self, subject):
-        if subject in self.subject_give_ins:
-            return self.subject_give_ins[subject]
-        return None
-
-    def __eq__(self, x):
-        return self.student_id == x.student_id
-
-
-
-
